move MarlinPresentation to presentationlayer package
